+++ src/scal/rest_server/api_v1/event_handlers_yearly.go
@@ -430,70 +430,70 @@ func UpdateYearly(req Request) (*Response, error) {
 			return nil, NewError(Internal, "", err)
 		}
 	}
-	// PARAM="timeZone", PARAM_TYPE="string"
+	// PARAM="timeZone", PARAM_TYPE="string", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.TimeZone,
 		lastEventModel.TimeZone,
 	) {
 		eventMeta.FieldsMtime["timeZone"] = now
 	}
-	// PARAM="timeZoneEnable", PARAM_TYPE="bool"
+	// PARAM="timeZoneEnable", PARAM_TYPE="bool", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.TimeZoneEnable,
 		lastEventModel.TimeZoneEnable,
 	) {
 		eventMeta.FieldsMtime["timeZoneEnable"] = now
 	}
-	// PARAM="calType", PARAM_TYPE="string"
+	// PARAM="calType", PARAM_TYPE="string", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.CalType,
 		lastEventModel.CalType,
 	) {
 		eventMeta.FieldsMtime["calType"] = now
 	}
-	// PARAM="summary", PARAM_TYPE="string"
+	// PARAM="summary", PARAM_TYPE="string", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.Summary,
 		lastEventModel.Summary,
 	) {
 		eventMeta.FieldsMtime["summary"] = now
 	}
-	// PARAM="description", PARAM_TYPE="string"
+	// PARAM="description", PARAM_TYPE="string", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.Description,
 		lastEventModel.Description,
 	) {
 		eventMeta.FieldsMtime["description"] = now
 	}
-	// PARAM="icon", PARAM_TYPE="string"
+	// PARAM="icon", PARAM_TYPE="string", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.Icon,
 		lastEventModel.Icon,
 	) {
 		eventMeta.FieldsMtime["icon"] = now
 	}
-	// PARAM="month", PARAM_TYPE="int"
+	// PARAM="month", PARAM_TYPE="uint8", PARAM_INT=true
 	if !reflect.DeepEqual(
 		eventModel.Month,
 		lastEventModel.Month,
 	) {
 		eventMeta.FieldsMtime["month"] = now
 	}
-	// PARAM="day", PARAM_TYPE="int"
+	// PARAM="day", PARAM_TYPE="uint8", PARAM_INT=true
 	if !reflect.DeepEqual(
 		eventModel.Day,
 		lastEventModel.Day,
 	) {
 		eventMeta.FieldsMtime["day"] = now
 	}
-	// PARAM="startYear", PARAM_TYPE="int"
+	// PARAM="startYear", PARAM_TYPE="int", PARAM_INT=true
 	if !reflect.DeepEqual(
 		eventModel.StartYear,
 		lastEventModel.StartYear,
 	) {
 		eventMeta.FieldsMtime["startYear"] = now
 	}
-	// PARAM="startYearEnable", PARAM_TYPE="bool"
+	// PARAM="startYearEnable", PARAM_TYPE="bool", PARAM_INT=false
 	if !reflect.DeepEqual(
 		eventModel.StartYearEnable,
 		lastEventModel.StartYearEnable,
@@ -670,7 +670,7 @@ func PatchYearly(req Request) (*Response, error) {
 	{
 		rawValue, ok := patchMap["month"]
 		if ok {
-			// json Unmarshal converts int to float64
+			// json Unmarshal converts uint8 to float64
 			value, typeOk := rawValue.(float64)
 			if !typeOk {
 				return nil, NewError(
@@ -679,7 +679,7 @@ func PatchYearly(req Request) (*Response, error) {
 					nil,
 				)
 			}
-			eventModel.Month = int(value)
+			eventModel.Month = uint8(value)
 			delete(patchMap, "month")
 			eventMeta.FieldsMtime["month"] = now
 		}
@@ -687,7 +687,7 @@ func PatchYearly(req Request) (*Response, error) {
 	{
 		rawValue, ok := patchMap["day"]
 		if ok {
-			// json Unmarshal converts int to float64
+			// json Unmarshal converts uint8 to float64
 			value, typeOk := rawValue.(float64)
 			if !typeOk {
 				return nil, NewError(
@@ -696,7 +696,7 @@ func PatchYearly(req Request) (*Response, error) {
 					nil,
 				)
 			}
-			eventModel.Day = int(value)
+			eventModel.Day = uint8(value)
 			delete(patchMap, "day")
 			eventMeta.FieldsMtime["day"] = now
 		}
@@ -882,7 +882,7 @@ func MergeYearly(req Request) (*Response, error) {
 	// B <== lastEventModel		<== The current (server's latest) data
 	now := time.Now()
 	func() {
-		// PARAM="timeZone", PARAM_TYPE="string"
+		// PARAM="timeZone", PARAM_TYPE="string", PARAM_INT=false
 		inputValue := inputEventModel.TimeZone
 		lastValue := lastEventModel.TimeZone
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -905,7 +905,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="timeZoneEnable", PARAM_TYPE="bool"
+		// PARAM="timeZoneEnable", PARAM_TYPE="bool", PARAM_INT=false
 		inputValue := inputEventModel.TimeZoneEnable
 		lastValue := lastEventModel.TimeZoneEnable
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -928,7 +928,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="calType", PARAM_TYPE="string"
+		// PARAM="calType", PARAM_TYPE="string", PARAM_INT=false
 		inputValue := inputEventModel.CalType
 		lastValue := lastEventModel.CalType
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -951,7 +951,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="summary", PARAM_TYPE="string"
+		// PARAM="summary", PARAM_TYPE="string", PARAM_INT=false
 		inputValue := inputEventModel.Summary
 		lastValue := lastEventModel.Summary
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -974,7 +974,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="description", PARAM_TYPE="string"
+		// PARAM="description", PARAM_TYPE="string", PARAM_INT=false
 		inputValue := inputEventModel.Description
 		lastValue := lastEventModel.Description
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -997,7 +997,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="icon", PARAM_TYPE="string"
+		// PARAM="icon", PARAM_TYPE="string", PARAM_INT=false
 		inputValue := inputEventModel.Icon
 		lastValue := lastEventModel.Icon
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -1020,7 +1020,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="month", PARAM_TYPE="int"
+		// PARAM="month", PARAM_TYPE="uint8", PARAM_INT=true
 		inputValue := inputEventModel.Month
 		lastValue := lastEventModel.Month
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -1043,7 +1043,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="day", PARAM_TYPE="int"
+		// PARAM="day", PARAM_TYPE="uint8", PARAM_INT=true
 		inputValue := inputEventModel.Day
 		lastValue := lastEventModel.Day
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -1066,7 +1066,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="startYear", PARAM_TYPE="int"
+		// PARAM="startYear", PARAM_TYPE="int", PARAM_INT=true
 		inputValue := inputEventModel.StartYear
 		lastValue := lastEventModel.StartYear
 		if reflect.DeepEqual(inputValue, lastValue) {
@@ -1089,7 +1089,7 @@ func MergeYearly(req Request) (*Response, error) {
 		}
 	}()
 	func() {
-		// PARAM="startYearEnable", PARAM_TYPE="bool"
+		// PARAM="startYearEnable", PARAM_TYPE="bool", PARAM_INT=false
 		inputValue := inputEventModel.StartYearEnable
 		lastValue := lastEventModel.StartYearEnable
 		if reflect.DeepEqual(inputValue, lastValue) {
